requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.24
Django==1.8.15
django-concurrency==1.3.2
django-extensions==1.7.4
kombu==3.0.37
pytz==2016.7
six==1.10.0
